fix: auto-fix golangci-lint v2 violations (QF/ST/misspell/whitespace)
Apply golangci-lint --fix to resolve ~1100 auto-fixable violations:
- QF1008: remove embedded field from selector (959)
- QF1012: merge if/else into switch (58)
- QF1003: convert if/else chain to switch (41)
- QF1001: apply De Morgan's law (17)
- ST1005: lowercase error strings (partial)
- misspell: fix typos (21)
- whitespace: remove trailing newlines (5)
- unconvert: remove unnecessary conversions (partial)
- Other QF/ST quick fixes

+++ internal/datacoord/import_util.go
@@ -335,7 +335,7 @@ func AssembleImportRequest(task ImportTask, job ImportJob, meta *meta, alloc all
 
 	idBegin, idEnd, err := common.AllocAutoID(func(n uint32) (int64, int64, error) {
 		ids, ide, e := alloc.AllocN(int64(n))
-		return int64(ids), int64(ide), e
+		return ids, ide, e
 	}, uint32(preAllocIDNum), Params.CommonCfg.ClusterID.GetAsUint64())
 	if err != nil {
 		return nil, err
@@ -829,7 +829,7 @@ func CalculateTaskSlot(task ImportTask, importMeta ImportMeta) int {
 	baseBufferSize := paramtable.Get().DataNodeCfg.ImportBaseBufferSize.GetAsInt()
 	if task.GetType() == ImportTaskType {
 		// ImportTask use dynamic buffer size calculated by vchannels and partitions
-		taskBufferSize = int(baseBufferSize) * len(job.GetVchannels()) * len(job.GetPartitionIDs())
+		taskBufferSize = baseBufferSize * len(job.GetVchannels()) * len(job.GetPartitionIDs())
 	} else {
 		// PreImportTask use fixed buffer size
 		taskBufferSize = baseBufferSize
